# src/docs/swagger.doc.json
{ 
  "openapi": "3.0.0",
  "info": {
    "title": "Workout Match",
    "description": "Workout Match is an API for creating and managing customized exercises and workouts. With this API, you can create exercises with sets, reps, and exercise classification (type), as well as group them into workouts that can have a defined difficulty level. Additionally, Workout Match also offers the ability to search for exercises and workouts from other users and copy them to your own account, which is especially useful for those looking for new workout ideas and variations.",
    "contact": {
      "email": "arthurgbsf@gmail.com"
    },
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Database"
    },
    {
      "url": "https://workout-match.onrender.com/api-docs",
      "description": "Production Server"
    }
  ],
  "paths": {
...
    "/workouts": {
      "get": {
        "parameters": [{
          "in":"query", 
          "name": "user", 
          "required": false,
          "schema":{"type": "boolean"},
        "description": "Optional. Pass 'user = true' to get the logged user workouts."}
        ],
        "description": "Route to get all users workouts registered. Workout created by user not appear in this request.",
        "tags": [
          "Workouts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Workouts"
                },
                "examples": {
                  "workouts": {
                    "$ref": "#/components/examples/workouts"
                  }
                }
              }
            }
          },
          "404": {
            "description": "There aren't workouts available. No user registered workouts.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "examples": {
                  "menssage": {
                    "value": {
                      "menssage": "No Workouts avaible."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/workouts/workout": {
      "post": {
        "description": "Route to create an workout.",
        "tags": [
          "Workouts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              },
              "examples": {
                "workout": {
                  "value": {
                    "workout": "Full Body Basic",
                    "level": "Intermediate",
                    "exercises": [
                      "644a87ae01fbb12d48cdc459",
                      "644a87ae01fbb12d48cdc45b",
                      "644a87ae01fbb12d48cdc45e"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The request was successful. The workout was created.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Message"
                },
                "examples": {
                  "menssage": {
                    "value": {
                      "menssage": "Workout created."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "As a prerequisite that there is at least one exercise in the array of exercises. Exercises also need to be created or copied by the user.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "examples": {
                  "error - exercise required": {
                    "$ref": "#/components/examples/validateExercisesError02"
                  },
                  "error - uncopied exercises": {
                    "$ref": "#/components/examples/validateExercisesError03"
                  }
                }
              }
            }
          },
          "404": {
            "description": "The user, exercises and workout are checked during the request.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "examples": {
                  "error - user not found": {
                    "$ref": "#/components/examples/getUserByIdAndCheck"
                  },
                  "error - exercises not found": {
                    "$ref": "#/components/examples/validateExercisesError01"
                  },
                  "error - workout not found": {
                    "$ref": "#/components/examples/setRefWorkoutInExercise"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/workouts/{id}": {
      "parameters": [
        {
          "in": "path",
          "name": "id",
          "schema": {
            "type": "string"
          },
          "required": true,
          "description": "The workout ID."
        }
      ],
      "get": {
        "description": "Route to get an workout by id.",
        "tags": [
          "Workouts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/Workout"
                  }
                },
                "examples": {
                  "workout": {
                    "$ref": "#/components/examples/workout"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400-ObjectIdCheck"
          },
          "404": {
            "$ref": "#/components/responses/404-getWorkoutByIdAndCheck"
          }
        }
      },
      "put": {
        "description": "Route to update an workout by Id. It is possible to add or remove exercises, including the id in the exercises array or removing the id.",
        "tags": [
          "Workouts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/workoutUpdateExamples"
        },
        "responses": {
          "400": {
            "description": "Id passed as parameter is invalid or training has not been updated. In the situation of updating exercises, errors related to the exercise may occur.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Message"
                },
                "examples": {
                  "error - invalid id": {
                    "$ref": "#/components/examples/invalid-id"
                  },
                  "error - not updated": {
                    "$ref": "#/components/examples/not-updated"
                  },
                  "error - exercise required": {
                    "$ref": "#/components/examples/validateExercisesError02"
                  },
                  "error - uncopied exercises": {
                    "$ref": "#/components/examples/validateExercisesError03"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401-ifParamIdIsUserIdCheck"
          },
          "404": {
            "description": "Exercises or Workout not found in database.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Message"
                },
                "examples": {
                  "error - exercise not found": {
                    "value": {
                      "message": "Exercises not found."
                    }
                  },
                  "error - workout not found": {
                    "value": {
                      "message": "Workout not found."
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "Route to delete an workout.",
        "tags": [
          "Workouts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Workout deleted with success.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "examples": {
                  "message": {
                    "value": {
                      "message": "The workout was deleted with success."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Id passed as parameter is invalid or the exercise wasn't deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "examples": {
                  "error - invalid id": {
                    "$ref": "#/components/examples/invalid-id"
                  },
                  "error - not deleted": {
                    "$ref": "#/components/examples/not-deleted"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401-ifParamIdIsUserIdCheck"
          },
          "404": {
            "$ref": "#/components/responses/404-getWorkoutByIdAndCheck"
          }
        }
      },
      "post": {
        "description": "Route to copy an workout. It can be the user's own or another user's.",
        "tags": [
          "Workouts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Copied with success.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "examples": {
                  "message": {
                    "value": {
                      "message": "Copied with success."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400-ObjectIdCheck"
          },
          "404": {
            "$ref": "#/components/responses/404-getWorkoutByIdAndCheck"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
...
      "Message": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "Exercise": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "exercise": {
            "type": "string"
          },
          "sets": {
            "type": "string"
          },
          "reps": {
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "copiedExerciseId": {
            "type": "string"
          },
          "inWorkouts": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "Workout-exercise-type-array-of-strings": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "workout": {
            "type": "string"
          },
          "level": {
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "copiedWorkoutId": {
            "type": "string"
          },
          "exercises": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "Workout": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "workout": {
            "type": "string"
          },
          "level": {
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "copiedWorkoutId": {
            "type": "string"
          },
          "exercises": {
            "type": "array",
            "items": {
              "type": "object",
              "$ref": "#/components/schemas/Exercise"
            }
          }
        }
      },
      "Workouts": {
        "type": "array",
        "items": {
          "type": "object",
          "$ref": "#/components/schemas/Workout"
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "400-ObjectIdCheck": {
        "description": "Id passed as parameter is invalid.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "$ref": "#/components/schemas/Message"
            },
            "examples": {
              "error": {
                "value": {
                  "message": "Invalid Id."
                }
              }
            }
          }
        }
      },
...
      "401-ifParamIdIsUserIdCheck": {
        "description": "Id passed as parameter not linked to the user.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object"
            },
            "examples": {
              "error": {
                "value": {
                  "message": "This id is not linked to this user."
                }
              }
            }
          }
        }
      },
      "404-getWorkoutByIdAndCheck": {
        "description": "Workout not found in database.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object"
            },
            "examples": {
              "error": {
                "value": {
                  "message": "Workout not found."
                }
              }
            }
          }
        }
      }
    },
    "requestBodies": {
...
      "workoutUpdateExamples": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Workout-exercise-type-array-of-strings"
            },
            "examples": {
              "update": {
                "value": {
                  "workout": "Full Body Basic",
                  "level": "Intermediate",
                  "exercises": [
                    "644a87ae01fbb12d48cdc459",
                    "644a87ae01fbb12d48cdc45b",
                    "644a87ae01fbb12d48cdc45e"
                  ]
                }
              },
              "workout-update": {
                "value": {
                  "workout": "Full Body"
                }
              },
              "level-update": {
                "value": {
                  "level": "Advanced"
                }
              },
              "exercise-update-add": {
                "value": {
                  "exercises": [
                    "644a87ae01fbb12d48cdc459",
                    "644a87ae01fbb12d48cdc45b",
                    "644a87ae01fbb12d48cdc45e",
                    "644a87ae01fbb12d48cdc45b"
                  ]
                }
              },
              "exercise-update-remove": {
                "value": {
                  "exercises": [
                    "644a87ae01fbb12d48cdc459",
                    "644a87ae01fbb12d48cdc45b"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "examples": {
      "invalid-id": {
        "value": {
          "message": "Invalid Id."
        }
      },
      "not-updated": {
        "value": {
          "message": "Wasn't updated."
        }
      },
      "not-deleted": {
        "value": {
          "message": "Wasn't deleted."
        }
      },
      "workouts": {
        "value": [
          {
            "_id": "644a7e1b0ff85088d51fcb5a",
            "workout": " Full Body Basic",
            "level": "beginner",
            "createdBy": "644a7d0d0ff85088d51fcb49",
            "createdAt": "27/04/2023 às 10:52:27",
            "exercises": [
              {
                "_id": "644a7d960ff85088d51fcb4f",
                "exercise": "Australian Pull ups",
                "sets": "3",
                "reps": "8",
                "type": "Pull",
                "createdAt": "27/04/2023 às 10:50:14"
              },
              {
                "_id": "644a7db70ff85088d51fcb52",
                "exercise": "Squats",
                "sets": "3",
                "reps": "8",
                "type": "Legs",
                "createdAt": "27/04/2023 às 10:50:47 "
              },
              {
                "_id": "644a8c03e716fd14d9a552b0",
                "exercise": "Push ups",
                "sets": "3",
                "reps": "8",
                "type": "Push",
                "createdAt": "27/04/2023 às 11:51:47 "
              }
            ],
            "updatedAt": "27/04/2023 às 13:23:17 "
          },
          {
            "_id": "644a7e1b0ff85088d51fcb5b",
            "workout": " Full Body Advanced",
            "level": "Advanced",
            "createdBy": "644a7d0d0ff85088d51fcb48",
            "createdAt": "27/04/2023 às 10:52:27",
            "exercises": [
              {
                "_id": "644a7d960ff85088d51fcb4g",
                "exercise": "Pull ups",
                "sets": "3",
                "reps": "8",
                "type": "Pull",
                "createdAt": "27/04/2023 às 10:50:14"
              },
              {
                "_id": "644a7db70ff85088d51fcb51",
                "exercise": "Air Squats",
                "sets": "3",
                "reps": "8",
                "type": "Legs",
                "createdAt": "27/04/2023 às 10:50:47 "
              },
              {
                "_id": "644a8c03e716fd14d9a552b2",
                "exercise": "Dips",
                "sets": "3",
                "reps": "8",
                "type": "Push",
                "createdAt": "27/04/2023 às 11:51:47 "
              }
            ],
            "updatedAt": "27/04/2023 às 13:23:17 "
          }
        ]
      },
...
      "getUserByIdAndCheck": {
        "value": {
          "message": "User not found."
        }
      },
      "validateExercisesError01": {
        "value": {
          "message": "Exercises not found."
        }
      },
      "validateExercisesError02": {
        "value": {
          "message": "Is required at least one exercise."
        }
      },
      "validateExercisesError03": {
        "value": {
          "message": "You need to copy theses exercises first : '644a7d690ff85088d51fcb4c'"
        }
      },
      "setRefWorkoutInExercise": {
        "value": {
          "message": "Workout not found."
        }
      },
      "workout": {
        "value": {
          "_id": "644a7e1b0ff85088d51fcb5a",
          "workout": " Full Body Basic",
          "level": "beginner",
          "createdBy": "644a7d0d0ff85088d51fcb49",
          "createdAt": "27/04/2023 às 10:52:27",
          "exercises": [
            {
              "_id": "644a7d960ff85088d51fcb4f",
              "exercise": "Australian Pull ups",
              "sets": "3",
              "reps": "8",
              "type": "Pull",
              "createdAt": "27/04/2023 às 10:50:14"
            },
            {
              "_id": "644a7db70ff85088d51fcb52",
              "exercise": "Squats",
              "sets": "3",
              "reps": "8",
              "type": "Legs",
              "createdAt": "27/04/2023 às 10:50:47 "
            },
            {
              "_id": "644a8c03e716fd14d9a552b0",
              "exercise": "Push ups",
              "sets": "3",
              "reps": "8",
              "type": "Push",
              "createdAt": "27/04/2023 às 11:51:47 "
            }
          ],
          "updatedAt": "27/04/2023 às 13:23:17 "
        }
      }
    }
  }
}